Refactor completion popup
